encapulate txplan notes into component
