+++ src/interface/src/app/services/notes.service.ts
@@ -17,7 +17,7 @@ export type NotesModelName = 'planning_area' | 'project_area';
   providedIn: 'root',
 })
 export abstract class BaseNotesService {
-  constructor(protected http: HttpClient) { }
+  constructor(protected http: HttpClient) {}
   protected abstract baseUrl: string;
   protected resource = '/note/';
