+++ src/interface/src/app/treatments/treatment-plan-notes/treatment-plan-notes.component.html
@@ -1,0 +1,8 @@
+<sg-notes-sidebar
+  [notes]="(notes$ | async) || []"
+  [notesState]="notesSidebarState"
+  (createNote)="addNote($event)"
+  (deleteNote)="handleNoteDelete($event)"
+  noNotesTitleText="No Treatment Plan Notes Yet"
+  noNotesDetailText="Start adding notes to help your team learn more about this treatment plan.">
+</sg-notes-sidebar>
