+++ src/interface/src/app/treatments/treatment-plan-tabs/treatment-plan-tabs.component.html
@@ -9,13 +9,6 @@
     <app-map-base-layer></app-map-base-layer>
   </mat-tab>
   <mat-tab label="Notes">
-    <sg-notes-sidebar
-      [notes]="(notes$ | async) || []"
-      [notesState]="notesSidebarState"
-      (createNote)="addNote($event)"
-      (deleteNote)="handleNoteDelete($event)"
-      noNotesTitleText="No Treatment Plan Notes Yet"
-      noNotesDetailText="Start adding notes to help your team learn more about this treatment plan.">
-    </sg-notes-sidebar>
+    <app-treatment-plan-notes></app-treatment-plan-notes>
   </mat-tab>
 </mat-tab-group>
